Fixed race condition in ezImageFormat registration (#1426)
The ezImageFormat class was used both as an ezEnumerable, as well as a local class. This creates a race condition and can result in a crash. Therefore the format type registration has been moved into a separate class.

Also added a couple of useful asserts to Jolt and improved the controller base class.