Fixed world reader to be able to instantiate prefabs from multiple threads (#1432)
The ezPrefabResource uses a single ezWorldReader to instantiate prefabs, no matter from which thread it is called.

Unfortunately, parts of its internal state are modified during instantiation, which means that if the same prefab gets instantiated in parallel (into different worlds), it would crash.

This can happen when you have two scenes open in the editor and then transform an asset that is used in both of them. Both scenes will update their prefab instances, which can happen on different threads which then use the same world reader.

The fix moves all mutable state into the instantiation context. I did not want to change the interface of ezWorldReader, which would have affected a lot of the code base. Therefore I had to use a thread_local to track the active instantiation context. ezWorldReader is already using thread_locals for the string deduplication context, so this doesn't change how it can be used.