Fix assert in Vulkan when closing scenes and wrong ctor reflection for ezBoxManipulatorAttribute (#1439)
It is unfortunately not possible to figure out the type of the Nth ctor parameter, so `EZ_CONSTRUCTOR_PROPERTY` relies on duck-typing to construct the object. This means that the `EZ_CONSTRUCTOR_PROPERTY` parameter type list can diverge from the actual types as long as the compiler can convert them. In this case, two ctor parameters were swapped but the reflection was not adapted which still worked because the compiler converted float to bool and vise verse to make it 'fit'.