Editor camera and selection improvements (#1441)
NOTE: This change may break your muscle-memory for how to move around a scene in the editor!

This PR introduces the major change that holding the left mouse button (LMB) and moving the mouse does NOT move the camera any longer. Unfortunately this did not work well with gizmos. If you slightly missed a gizmo handle and clicked next to it, it was easy to move the camera far away, interrupting your workflow. With this change, in such a situation nothing happens anymore .

To move the camera the way LMB did before, instead hold CTRL and use the middle mouse button (MMB) to drag. This is analogous to MMB which already moved the camera in the ground plane, now CTRL+MMB is just a variation of that.

Additionally, the orbit camera was changed, so that the orbit reference point is always right in front of the camera, just at varying distance. Thus using orbit mode (hold ALT) never makes the camera jump to some other point anymore, which was often very annoying. Instead the orbit reference point moves with the camera when it rotates or moves.

Other changes:
* When holding the ALT key, all gizmos now disappear, to indicate that ALT activates the orbit camera.
* Made left-click to select a bit more lenient. If you slightly move the mouse, it still counts as a selection click.
* The SkyLight component is now by default set to 'Static', to not waste performance in cases where all you want is prettier ambient lighting.