Make object creation with a desired component type easier (#1446)
When creating an empty game object somewhere (CTRL+SHIFT+X in viewport), the editor now shows the component type menu. If the user choses a component type from it, the new game object uses that. If they click somewhere else to close the menu, it uses the default shape icon component, as before.