C++ Project Generation now also builds stand-alone game exe (#1465)
The C++ project generation now not only builds the game plugin, but also a stand-alone binary that is very similar to ezPlayer, just even more trimmed down, to immediately play your game with.

This works directly from the Visual Studio solution and the binary can also be used in an exported game (though it first has to be whitelisted to be exported, at all).

This will probably get more tightly integrated with the export step in the future.

Also moved the RTS and asteroids sample to use the C++ project generation infrastructure, instead of a separate engine plugin.