Made type accessor convenience functions accessible in derived classes + other fixes (#1473)
The type accessors used function overloads to add convenience functions. Derived classes would have to implement the virtual interface, which unfortunately in C++ hides the base class functions. Thus all convenience functions were inaccessible, unless you got a pointer to the base class (which we did in some cases).

I renamed the functions. There are other solutions in C++ to achieve the same, but they are error prone. With the new names it is also a bit easier to see the differences.

I've also fixed a couple of other issues, For example you could compare a hashed string with a constant string, but it would not do the correct comparison.